server/requirements.txt
flask
flask-sqlalchemy
flask-cors
requests
beautifulsoup4
google-generativeai
gunicorn